Move checkBrowserVersion into firefox subfolder
We're moving to a plugin architecture, where each "target" is it's own
module. Right now we only have firefox as a target, but we plan to add
others. Before we start doing that I wanted to prepare a little by
moving files that are firefox-specific into a firefox folder. That way
it should be easier to detach it as a target it later on.
