+++ src/server/firefox/__tests__/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  env: {
+    jasmine: true,
+    jest: true,
+  },
+};
